requirements.txt
pyserial>=3.5
requests>=2.28.0
weasyprint>=60.0





